requirements.txt
pillow
werkzeug
numpy
